# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Open Wallet API.",
        "title": "Open Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/recurrences/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a recurrence template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurrences"
                ],
                "summary": "Delete Recurrence By ID",
                "operationId": "v1DeleteRecurrence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recurrence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "all",
                            "until_current"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "Handling of linked transactions: 'all' (default) deletes the recurrence and all related transactions (past/future); 'until_current' preserves past history but removes future recurrences.",
                        "name": "scope",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recurrence deleted"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a recurrence template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurrences"
                ],
                "summary": "Update a recurrence",
                "operationId": "v1UpdateRecurrence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recurrence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recurrence payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_resources_recurrences_handlers.UpdateRecurrenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurrence updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_recurrences_handlers_UpdateRecurrenceResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/recurrences/{period}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates entry records for all recurrence templates that don't already have one in the given period.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurrences"
                ],
                "summary": "Prepare recurrences for a period",
                "operationId": "v1PrepareRecurrence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period in YYYYMM format (e.g. 202603)",
                        "name": "period",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recurrences prepared"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/transactions/{transaction_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a transaction and all entries related by the ID of the transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Delete Transaction By ID",
                "operationId": "v1DeleteTransaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction ID",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction deleted"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Update a transaction",
                "operationId": "v1UpdateTransaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction ID",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Installment payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_resources_transactions_handlers.UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Installment updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_transactions_handlers_UpdateTransactionResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_recurrences_handlers_UpdateRecurrenceResponseData": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_recurrences_handlers.UpdateRecurrenceResponseData"
                }
            }
        },
...
        "github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_transactions_handlers_UpdateTransactionResponseData": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_transactions_handlers.UpdateTransactionResponseData"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPErrorData"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util_httputil.HTTPErrorData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_resources_recurrences_handlers.RecurrenceResource": {
            "type": "object",
            "required": [
                "amount",
                "created_at",
                "day_of_month",
                "id",
                "name",
                "start_period",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_color": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "day_of_month": {
                    "type": "integer"
                },
                "end_period": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "start_period": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_resources_recurrences_handlers.UpdateRecurrenceRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "string"
                },
                "day_of_month": {
                    "type": "integer",
                    "maximum": 31,
                    "minimum": 1
                },
                "end_period": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "note": {
                    "type": "string",
                    "maxLength": 400,
                    "minLength": 0
                },
                "start_period": {
                    "type": "string"
                }
            }
        },
        "internal_resources_recurrences_handlers.UpdateRecurrenceResponseData": {
            "type": "object",
            "required": [
                "recurrence"
            ],
            "properties": {
                "recurrence": {
                    "$ref": "#/definitions/internal_resources_recurrences_handlers.RecurrenceResource"
                }
            }
        },
...
        "internal_resources_transactions_handlers.TransactionResource": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "name",
                "type",
                "user_id"
            ],
            "properties": {
                "category_color": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "recurrence_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_resources_transactions_handlers.UpdateEntryRequest": {
            "type": "object",
            "required": [
                "amount",
                "reference_date"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "maximum": 999999,
                    "minimum": -999999
                },
                "reference_date": {
                    "type": "string"
                }
            }
        },
        "internal_resources_transactions_handlers.UpdateTransactionRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/internal_resources_transactions_handlers.UpdateEntryRequest"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "note": {
                    "type": "string",
                    "maxLength": 400,
                    "minLength": 0
                }
            }
        },
        "internal_resources_transactions_handlers.UpdateTransactionResponseData": {
            "type": "object",
            "required": [
                "transaction"
            ],
            "properties": {
                "transaction": {
                    "$ref": "#/definitions/internal_resources_transactions_handlers.TransactionResource"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}